Add map colour style function to assets
